introduced navigate class (.hpp and .cpp) and navigator object in nodo_destino
