+++ practica4/src/practica4/navigate.cpp
@@ -1,0 +1,32 @@
+#include "practica4/navigate.hpp"
+
+#include <ros/ros.h>
+#include <move_base_msgs/MoveBaseAction.h>
+#include <actionlib/client/simple_action_client.h>
+
+namespace practica4
+{
+
+void doneCb(const actionlib::SimpleClientGoalState& state,
+  const move_base_msgs::MoveBaseResultConstPtr& result)
+{
+  ROS_INFO("Terminado!!");
+  finished_ = true;
+}
+
+void feedbackCb(const move_base_msgs::MoveBaseFeedbackConstPtr& feedback)
+{
+  double goal_x = goal.target_pose.pose.position.x;
+  double goal_y = goal.target_pose.pose.position.y;
+  double current_x = feedback->base_position.pose.position.x;
+  double current_y = feedback->base_position.pose.position.y;
+
+  double diff_x = goal_x - current_x;
+  double diff_y = goal_y - current_y;
+
+  double dist = sqrt(diff_x * diff_x + diff_y * diff_y);
+
+  ROS_INFO("Distance to goal = %lf", dist);
+}
+
+}//practica4
